+++ src/index.ts
@@ -46,8 +46,10 @@ type Options = {
   plugins?: [string, Record<string, any>][];
 };
 
+const isWebContainer = globalThis.process?.versions?.webcontainer;
+
 const react = (_options?: Options): PluginOption[] => {
-  let hmrDisabled = true;
+  let hmrDisabled = false;
   const options = {
     jsxImportSource: _options?.jsxImportSource ?? "react",
     tsDecorators: _options?.tsDecorators,
@@ -88,6 +90,11 @@ const react = (_options?: Options): PluginOption[] => {
         ) {
           throw new Error(
             "[vite:react-swc] The MDX plugin should be placed before this plugin",
+          );
+        }
+        if (isWebContainer) {
+          config.logger.warn(
+            "[vite:react-swc] SWC is currently not supported in WebContainers. You can use the default React plugin instead.",
           );
         }
       },