+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Disable Fast Refresh based on `config.server.hmr === false` instead of `process.env.TEST`
+- Warn when plugin is in WebContainers (see [#118](https://github.com/vitejs/vite-plugin-react-swc/issues/118))
 
 ## 3.3.2